refactor: move logic to the todo component
