+++ apps/angular/5-crud-application/src/app/todo.model.ts
