+++ apps/angular/5-crud-application/src/index.html
@@ -15,6 +15,6 @@
       rel="stylesheet" />
   </head>
   <body>
-    <app-root></app-root>
+    <app-todo></app-todo>
   </body>
 </html>
