+++ apps/angular/5-crud-application/src/main.ts
@@ -2,9 +2,9 @@ import { provideZoneChangeDetection } from '@angular/core';
 import { appConfig } from './app/app.config';
 
 import { bootstrapApplication } from '@angular/platform-browser';
-import { AppComponent } from './app/app.component';
+import { TodoComponent } from './app/todo.component';
 
-bootstrapApplication(AppComponent, {
+bootstrapApplication(TodoComponent, {
   ...appConfig,
   providers: [provideZoneChangeDetection(), ...appConfig.providers],
 }).catch((err) => console.error(err));
